requirements.txt
fastapi
uvicorn
python-multipart
streamlit      
langchain
langchain-community
tavily-python
sentence-transformers
faiss-cpu
rank_bm25
python-dotenv
python-docx
pdfplumber
groq
pyyaml
langgraph
opik
pandas
numpy
litellm